feat: minor cleanup removing fonts and terminals
I am trying to use "more default" font and apps, thus I removed fonts
(Fira Code and Hack) and terminals (Alacritty and Kitty).

As the world is becoming more Wayland, I can use foot terminal reliably
even if on Gnome or KDE, so that was added, but is the default on Sericea.